backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.40.68",
    "fastapi>=0.121.0",
    "fpdf>=1.7.2",
    "mediapipe>=0.10.14",
    "numpy>=2.3.4",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "pyannote-audio>=3.4.0",
    "python-multipart>=0.0.20",
    "soundfile>=0.13.1",
    "uvicorn[standard]>=0.38.0",
    "websockets>=15.0.1",
]
